Added lecs and slides

+++ Lec8/basic_to_basic_tc.cpp
@@ -3,8 +3,8 @@ using namespace std;
 
 int main(){
     double height = 160.55;
-    int height_m = height/100;
-    int height_cms = height-height_m*100;
+    int height_m = (int)height/100; // Explicit
+    int height_cms = height-height_m*100; // Implicit
 
     cout<<height_m<<" mts and "<<height_cms<<" cms."<<endl;
